+++ tests/reduct/server_api_test.cc
@@ -72,13 +72,13 @@ TEST_CASE("reduct::Client should return error", "[server_api]") {
   auto client = IClient::Build("http://127.0.0.1:99999");
   auto [info, err] = client->GetInfo();
 
-  REQUIRE(err == Error{.code = -1, .message = "Could not establish connection"});
+  REQUIRE(err.code == -1);
 }
 
 TEST_CASE("reduct::Client should return 404 if base path wrong") {
   auto client = IClient::Build("http://127.0.0.1:8383/wrong_path");
   auto [info, err] = client->GetInfo();
-  REQUIRE(err == Error{.code = 404, .message = "Not found"});
+  REQUIRE(err.code == 404);
 }
 
 TEST_CASE("reduct::Client should return current token name and permissions", "[server_api][token_api]") {